requirements.txt
oauth2client==4.1.3
gspread==5.7.2
langchain==0.1.12
scipy==1.10.1
beautifulsoup4==4.11.1
faiss-cpu==1.7.2
fpdf==1.7.2
google-api-python-client==2.52.0
google-auth==2.15.0
google-auth-oauthlib==0.5.3
langchain-groq==0.1.0
matplotlib==3.7.1
numpy==1.24.3
pandas==1.4.1
PyPDF2==3.0.1
python-docx==0.8.11
python-dotenv==0.20.0
requests==2.28.1
seaborn==0.11.2
streamlit==1.20.0
transformers==4.26.0